move redis limit test
