+++ tests/Mutex/RedisMutexWithPredisTest.php
@@ -13,7 +13,7 @@ use PHPUnit\Framework\TestCase;
 use Predis\ClientInterface as PredisClientInterface;
 use Predis\PredisException;
 
-interface PredisClientInterfaceWithSetAndEvalMethods extends PredisClientInterface
+interface PredisClientInterfaceWithSetAndEvalMethods2 extends PredisClientInterface
 {
     /**
      * @return mixed
@@ -39,7 +39,7 @@ class RedisMutexWithPredisTest extends TestCase
     {
         parent::setUp();
 
-        $this->client = $this->createMock(PredisClientInterfaceWithSetAndEvalMethods::class);
+        $this->client = $this->createMock(PredisClientInterfaceWithSetAndEvalMethods2::class);
 
         $this->mutex = new RedisMutex($this->client, 'test', 2.5, 3.5);
     }
@@ -96,23 +96,6 @@ class RedisMutexWithPredisTest extends TestCase
         });
 
         self::assertTrue($executed);
-    }
-
-    public function testAcquireExpireTimeoutLimit(): void
-    {
-        $this->mutex = new RedisMutex($this->client, 'test');
-
-        $this->client->expects(self::once())
-            ->method('set')
-            ->with('php-malkusch-lock:test', new IsType(IsType::TYPE_STRING), 'PX', 31_557_600_000_000, 'NX')
-            ->willReturnSelf();
-
-        $this->client->expects(self::once())
-            ->method('eval')
-            ->with(self::anything(), 1, 'php-malkusch-lock:test', new IsType(IsType::TYPE_STRING))
-            ->willReturn(true);
-
-        $this->mutex->synchronized(static function () {});
     }
 
     /**
